migration to Java 17, Spring Boot 3.X and Maven
