+++ TourGuide/src/main/java/com/openclassrooms/tourguide/helper/InternalTestHelper.java
@@ -1,0 +1,15 @@
+package com.openclassrooms.tourguide.helper;
+
+public class InternalTestHelper {
+
+	// Set this default up to 100,000 for testing
+	private static int internalUserNumber = 100;
+	
+	public static void setInternalUserNumber(int internalUserNumber) {
+		InternalTestHelper.internalUserNumber = internalUserNumber;
+	}
+	
+	public static int getInternalUserNumber() {
+		return internalUserNumber;
+	}
+}
