+++ TourGuide/src/main/java/com/openclassrooms/tourguide/user/UserPreferences.java
@@ -1,0 +1,54 @@
+package com.openclassrooms.tourguide.user;
+
+public class UserPreferences {
+	
+	private int attractionProximity = Integer.MAX_VALUE;
+	private int tripDuration = 1;
+	private int ticketQuantity = 1;
+	private int numberOfAdults = 1;
+	private int numberOfChildren = 0;
+	
+	public UserPreferences() {
+	}
+	
+	public void setAttractionProximity(int attractionProximity) {
+		this.attractionProximity = attractionProximity;
+	}
+	
+	public int getAttractionProximity() {
+		return attractionProximity;
+	}
+	
+	public int getTripDuration() {
+		return tripDuration;
+	}
+
+	public void setTripDuration(int tripDuration) {
+		this.tripDuration = tripDuration;
+	}
+
+	public int getTicketQuantity() {
+		return ticketQuantity;
+	}
+
+	public void setTicketQuantity(int ticketQuantity) {
+		this.ticketQuantity = ticketQuantity;
+	}
+	
+	public int getNumberOfAdults() {
+		return numberOfAdults;
+	}
+
+	public void setNumberOfAdults(int numberOfAdults) {
+		this.numberOfAdults = numberOfAdults;
+	}
+
+	public int getNumberOfChildren() {
+		return numberOfChildren;
+	}
+
+	public void setNumberOfChildren(int numberOfChildren) {
+		this.numberOfChildren = numberOfChildren;
+	}
+
+}
