+++ TourGuide/src/main/resources/application.properties
@@ -1,2 +1,1 @@
-
-logging.level.tourGuide=DEBUG
+logging.level.com.openclassrooms.tourguide=DEBUG
